+++ src/services/customer.service.ts
@@ -1,7 +1,7 @@
 import axios from "axios";
 import Customer from "../models/customer.model";
 
-const baseUrl = "http://192.168.132.199:3001"
+const baseUrl = "http://192.168.225.128:3001"
 
 class CustomerService {
 
@@ -9,8 +9,20 @@ class CustomerService {
     return axios.get(baseUrl + "/api/v1/customers");
   }
 
-  create(data: Customer) {
+  create(data) {
     return axios.post(baseUrl + "/api/v1/customers", data)
+  }
+
+  getById(id) {
+    return axios.get(baseUrl + "/api/v1/customer/" + id)
+  }
+
+  update(id, data) {
+    return axios.put(baseUrl + `/api/v1/customer/${id}`, data)
+  }
+
+  delete(id) {
+    return axios.delete(baseUrl + "/api/v1/customer/" +  id)
   }
 
 }
